requirements.txt
Flask
flask-restx
psycopg2-binary 
python-dotenv